Module CodeUnit.CodeUnit

type t

Some metadata about a given code unit.

val origin : Global.t -> id

Return the name of the code unit that a symbol originated from.

val id : t -> id

The name of a given code unit

val imports : t -> id list

All of the code unit this unit directly imports

val create : id -> t

Create a code unit.

val add_global : Ident.t -> Domain.tp -> Domain.con option -> t -> Global.t * t

Add a binding to a given code unit.

val resolve_global : Ident.t -> t -> Global.t option

Attempt to resolve an identifier a given code unit.

val get_global : Global.t -> t -> Domain.tp * Domain.con option

Get the binding associated with a symbol.

val add_import : [< `Print of string option ] Yuujinchou.Pattern.t -> t -> t -> t

Add another code unit as an import.