Module DomainData.Make

Parameters

module Symbol : Basis.Symbol.S

Signature

module S : sig ... end

type dim = Cubical.Dim.t

type cof = (dim, int) Cubical.Cof.cof

type 'a stable_code = [

| `Pi of 'a * 'a	(*
Dependent product type
*)

| `Sg of 'a * 'a	(*
Dependent sum type
*)

| `Signature of (Ident.user * 'a) list	(*
First-Class Record types
*)

| `Ext of int * 'a * [ `Global of 'a ] * 'a	(*
Extension type
*)

| `Nat	(*
Natural numbers type
*)

| `Circle	(*
The circle S1.
*)

| `Univ	(*
A code for the universe (antinomous for now).
*)

]

A type code whose head constructor is stable under dimension substitution.

type 'a unstable_code = [

| `HCom of dim * dim * cof * 'a	(*
Formal composite types
*)

| `V of dim * 'a * 'a * 'a	(*
V types, for univalence
*)

]

A type code which may or may not be stable under dimension substitution. That is, type codes with these constructors may or may not remain in normal forms under substitution.

type env = {

tpenv : tp Basis.Bwd.bwd;

conenv : con Basis.Bwd.bwd;

}

and 'a clo =

| Clo of 'a * env

A closure combines a semantic environment with a syntactic object binding an additional variable.

and tp_clo = S.tp clo

and tm_clo = S.t clo

and sign_clo = S.sign clo

and con =

| Lam of Ident.t * tm_clo

| BindSym of Cubical.DimProbe.t * con	(*
A nominal binder of a dimension; these are used during the execution of coercion, which must probe a line of type codes with a fresh dimension.
*)

| LetSym of dim * Cubical.DimProbe.t * con	(*
An explicit substitution of a dimension for a symbol.
*)

| Cut of {

tp : tp;

cut : cut;

}	(*
Our notion of neutral value, a type annotated cut.
*)

| Zero

| Suc of con

| Base

| Loop of dim

| Pair of con * con

| Struct of (Ident.user * con) list

| SubIn of con

| ElIn of con	(*
The introduction form for the extension of a stable type code only (see ElStable and ElUnstable).
*)

| Dim0

| Dim1

| DimProbe of Cubical.DimProbe.t

| Cof of (con, con) Cubical.Cof.cof_f	(*
A mixin of the language of cofibrations (as described in Cubical.Cof), with dimensions and indeterminates in con.
*)

| Prf

| FHCom of [ `Nat | `Circle ] * dim * dim * cof * con

| StableCode of con stable_code

| UnstableCode of con unstable_code

| Box of dim * dim * cof * con * con

| VIn of dim * con * con * con

| Split of (cof * tm_clo) list

| LockedPrfIn of con

Value constructors are governed by con; we do not maintain in the datatype a priori any invariant that these represent whnfs (weak head normal forms). Whether a value constructor is a whnf is contingent on the ambient local state, such as the cofibration theory.

and tp =

| Sub of tp * cof * tm_clo

| Univ

| ElCut of cut

| ElStable of con stable_code

| ElUnstable of con unstable_code

| TpDim

| TpCof

| TpPrf of cof

| TpSplit of (cof * tp_clo) list

| Pi of tp * Ident.t * tp_clo

| Sg of tp * Ident.t * tp_clo

| Signature of sign

| Nat

| Circle

| TpLockedPrf of cof

and sign =

| Field of Ident.user * tp * S.sign clo

| Empty

and hd =

| Global of Symbol.t	(*
A top-level declaration
*)

| Var of int	(*
De Bruijn level
*)

| Coe of con * dim * dim * con

| UnstableCut of cut * unstable_frm

A head is a variable (e.g. Global, Var), or it is some kind of unstable elimination form (Coe, UnstableCut). The geometry of cut, hd, unstable_frm enables a very direct way to re-reduce a complex cut to whnf by following the unstable nodes to the root.

and cut = hd * frm list

A cut is a value that is blocked on the computation of a hd ("head"); when the head is computed, the list of stack frames (frm) carried by the cut will be enacted.

and frm =

| KAp of tp * con

| KFst

| KSnd

| KProj of Ident.user

| KNatElim of con * con * con

| KCircleElim of con * con * con

| KElOut	(*
The elimination form for the extension of a stable type code only (see ElStable).
*)

A stable frame is a dimension substitution-stable elimination form with a hole in place of its principal argument. Unstable elimination forms are governed by hd to ease the "re-reduction" of a value to whnf under a stronger cofibration theory.

and unstable_frm =

| KHCom of dim * dim * cof * con

| KCap of dim * dim * cof * con

| KVProj of dim * con * con * con

| KSubOut of cof * tm_clo

| KLockedPrfUnlock of tp * cof * con

An unstable frame is a dimension substitution-unstable elimination form with a hole in place of its principal argument.