Module Syntax.Make

Parameters

module Symbol : Basis.Symbol.S

Signature

module Cof = Kado.Syntax.Endo

type t =

| Var of int

| Global of Symbol.t

| Let of t * Ident.t * t

| Ann of t * tp

| Zero

| Suc of t

| NatElim of t * t * t * t

| Base

| Loop of t

| CircleElim of t * t * t * t

| Lam of Ident.t * t

| Ap of t * t

| Pair of t * t

| Fst of t

| Snd of t

| Struct of (Ident.user * t) list

| Proj of t * Ident.user

| Coe of t * t * t * t

| HCom of t * t * t * t * t

| Com of t * t * t * t * t

| SubIn of t

| SubOut of t

| Dim0

| Dim1

| Cof of ( t, t ) Cof.t

| ForallCof of t

| CofSplit of (t * t) list

| Prf

| ElIn of t

| ElOut of t

| Box of t * t * t * t * t

| Cap of t * t * t * t * t

| VIn of t * t * t * t

| VProj of t * t * t * t * t

| CodeExt of int * t * [ `Global of t ] * t

| CodePi of t * t

| CodeSg of t * t

| CodeSignature of (Ident.user * t) list

| CodeNat

| CodeUniv

| CodeV of t * t * t * t

| CodeCircle

| ESub of sub * t

and tp =

| Univ

| El of t

| TpVar of int

| TpDim

| TpCof

| TpPrf of t

| TpCofSplit of (t * tp) list

| Sub of tp * t * t

| Pi of tp * Ident.t * tp

| Sg of tp * Ident.t * tp

| Signature of sign

| Nat

| Circle

| TpESub of sub * tp

and sign = (Ident.user * tp) list

and sub =

| SbI

| SbC of sub * sub

| Sb1

| SbE of sub * t

| SbP

module CofBuilder : sig ... end

val tm_abort : t

Convenience constructors

val tp_abort : tp

Pretty printers

For display

These pretty printers are context sensitive, in order to recall user-specified names for binders.

val pp : Basis.Pp.env -> t Basis.Pp.printer

Print a core language term.

val pp_sign : Basis.Pp.env -> sign Basis.Pp.printer

Print a signature.

val pp_tp : Basis.Pp.env -> tp Basis.Pp.printer

Print a core language type.

val pp_sequent : lbl:string option -> (Ident.t * tp) list -> tp Basis.Pp.printer

Vertically print an iterated dependent product type as if it were a sequent, for display of goals.

val pp_partial_sequent : [< `BdrySat | `BdryUnsat ] -> (Ident.t * tp) list -> (t * tp) Basis.Pp.printer

Vertically print an iterated dependent product type as if it were a sequent, for display of goals. This variant will also print out a partially constructed terms, as well as display if the boundary conditions are met.

For debugging

When debugging, we are not likely to have enough context to use the nice pretty printers above; as a last resort, dump and dump_tp may be used.

val dump : t Basis.Pp.printer

val dump_sign : sign Basis.Pp.printer

val dump_tp : tp Basis.Pp.printer