Module Core__Conversion

The purpose of this module is to check whether two well-typed objects are equal or not. The semantics are that all definitions are unfolded, regardless of the "veil", since definitional equality is closed under unfolding of definitions.

module D := Core.Domain

module Error : sig ... end

exception ConversionError of Error.t

val equate_con : D.tp -> D.con -> D.con -> unit Core.Monads.conversion

val equate_tp : D.tp -> D.tp -> unit Core.Monads.conversion