Module Core__Quote

The purpose of this module is to transform semantic objects into syntactic objects as efficiently as possible; only user-specified top-level definitions will be unfolded, in accordance with the veil (see Monads.QuM.read_veil).

module D := Core.Domain

module S := Core.Syntax

val quote_con : D.tp -> D.con -> S.t Core.Monads.quote

val quote_tp : D.tp -> S.tp Core.Monads.quote

val quote_cut : D.cut -> S.t Core.Monads.quote

val quote_cof : D.cof -> S.t Core.Monads.quote

val quote_dim : D.dim -> S.t Core.Monads.quote