Module Core__Refiner

This is the basis of trusted inference rules for cooltt. This module also contains some auxiliary tactics, but these don't need to be trusted so they should be moved elsewhere.

module D := Core.Domain

module S := Core.Syntax

module EM := Core.Monads.ElabM

module CS := Core.ConcreteSyntax

type ('a, 'b) quantifier = 'a -> (Core.Ident.t * (Core.Tactic.var -> 'b)) -> 'b

module Hole : sig ... end

module Goal : sig ... end

module Dim : sig ... end

module Cof : sig ... end

module Prf : sig ... end

module Univ : sig ... end

module El : sig ... end

module ElV : sig ... end

module ElHCom : sig ... end

module Pi : sig ... end

module Sg : sig ... end

module Sub : sig ... end

module Nat : sig ... end

module Circle : sig ... end

module Structural : sig ... end

module Tactic : sig ... end