Module Cubical.Cof

The abstract syntax of the restricted predicate logic of cofibrations.

Types

type ('r, 'a) cof_f =

| Eq of 'r * 'r

| Join of 'a list

| Meet of 'a list

Multiple types in cooltt will need to include the langauge of cofibrations, relative to a particular interval algebra 'r. Therefore, we define a family polynomial endofunctors ('r, -) cof_f indexed in an interpretation of the interval algebra 'r.

type ('r, 'v) cof =

| Cof of ('r,Â ('r,Â 'v) cof) cof_f

| Var of 'v

For each interval algebra 'r, we define the free monad ('r, -) cof on the polynomial endofunctor ('r, -) cof_f: each ('r, 'v) cof is the language of cofibrations over an interval algebra 'r, with indeterminates drawn from 'v.

Smart constructors

val var : 'v -> ('a,Â 'v) cof

The cofibration corresponding to a variable Ï† : ğ”½.

val eq : 'a -> 'a -> ('a,Â 'v) cof

Given dimensions r, r' : ğ•€, the cofibration r=r'.

val join : ('a,Â 'v) cof list -> ('a,Â 'v) cof

Given a list Ï†0,...,Ï†n : ğ”½ of cofibrations, the disjunction Ï†0 âˆ¨ ... âˆ¨ Ï†n.

val meet : ('a,Â 'v) cof list -> ('a,Â 'v) cof

Given a list Ï†0,...,Ï†n : ğ”½ of cofibrations, the conjunction Ï†0 âˆ ... âˆ Ï†n.

val bot : ('a,Â 'v) cof

The false cofibration, equivalent to join [].

val top : ('a,Â 'v) cof

The true cofibration, equivalent to meet [].

val boundary : Dim.dim -> (Dim.dim,Â 'v) cof

The boundary âˆ‚r of a dimension r : ğ•€ is the disjunction r=0 âˆ¨ r=1