Module Cubical.CofThy

The CofThy module implements decision procedures for sequents relative to a theory over the interval, stated in the language of cofibrations.

type cof = Cubical__.CofThyData.cof

module Alg = CofThy.Alg

Algebraic theories over the interval.

module Disj = CofThy.Disj

Disjunctive theories over the interval.