Module CofThy.Alg

type t = alg_thy

type t' = alg_thy'

val emp' : alg_thy'

val empty : [> `Consistent of alg_thy' ]

val consistency : [< `Consistent of 'a | `Inconsistent ] -> [> `Consistent | `Inconsistent ]

val assume_vars : t' -> VarSet.t -> alg_thy'

val test_eq : t' -> (UF.key * UF.key) -> bool

val unsafe_test_and_assume_eq : t' -> (UF.key * UF.key) -> bool * alg_thy'

unsafe_test_and_assume_eq fuses test_eq and assume_eq (if there was one). * It is "unsafe" because we do not check consistency here.

val test_eqs : t' -> (UF.key * UF.key) list -> bool

val test_var : t' -> VarSet.elt -> bool

val test_vars : t' -> VarSet.t -> bool

val test_branch : t' -> (VarSet.t * (UF.key * UF.key) list) -> bool

val reduce_vars : t' -> VarSet.t -> VarSet.t

reduced_vars takes out redundant cofibration variables.

val reduce_eqs : t' -> (UF.key * UF.key) list -> [> `Consistent of t' * (UF.key * UF.key) list | `Inconsistent ]

reduce_eqs detects inconsistency of an equation set and takes out * redundant equations.

val reduce_branch : t' -> (VarSet.t * (UF.key * UF.key) list) -> [> `Consistent of alg_thy' * (VarSet.t * (UF.key * UF.key) list) | `Inconsistent ]

reduce_branch detects inconsistency of a branch and takes out redundant * cofibration variables and equations.

val reduce_branches : t' -> (VarSet.t * eq list) list -> cached_branches

reduce_branches removes inconsistent branches and takes out redundant * cofibration variables and equations.

val drop_useless_branches : (t' * branch) list -> cached_branches

drop_useless_branches drops the branches that could be dropped without * affecting the coverages.

val split : t -> cof list -> t list

split combines all the optimizers above to split an algebraic theory * into multiple ones induced by the input cofibration context.

val test : alg_thy' -> (Cubical.Dim.dim, int) Cubical.Cof.cof -> bool

test checks whether a cofibration is true within an algebraic theory

val left_invert_under_cofs : zero:'a -> seq:((t -> 'a) -> t list -> 'a) -> t -> cof list -> (t -> 'a) -> 'a