Module CofThy.Disj

Disjunctive theories over the interval.

type t

The type of a disjunctive theory.

val empty : t

The empty theory.

val envelope_alg : Alg.t -> t

Construct the enveloping disjunctive theory of an algebraic theory.

val consistency : t -> [ `Consistent | `Inconsistent ]

Returns the consistency of the theory.

val assume : t -> cof list -> t

Assumes the truth of a cofibration.

val test_sequent : t -> cof list -> cof -> bool

Tests the validity of a sequent against the supplied theory. Equivalent to assuming the conjunction of the context and then testing truth.

val left_invert : zero:'a -> seq:((Alg.t -> 'a) -> Alg.t list -> 'a) -> t -> (Alg.t -> 'a) -> 'a

Search all branches induced by irreducible joins under additional cofibrations.