Module Cubical__CofThy

module CofVar : sig ... end

type cof = (Cubical.Dim.dim, CofVar.t) Cubical.Cof.cof

module UF : sig ... end

module VarSet : sig ... end

type alg_thy' = {

classes : UF.t;
equivalence classes of dimensions

true_vars : VarSet.t;

}

A presentation of an algebraic theory over the language of intervals and cofibrations.

type eq = Cubical.Dim.dim * Cubical.Dim.dim

type branch = VarSet.t * eq list

A branch represents the meet of a bunch of atomic cofibrations.

type branches = branch list

type cached_branch = alg_thy' * branch

A cached_branch is a branch together with an algebraic theory * representing the resulting theory at the end of the branch.

type cached_branches = cached_branch list

type alg_thy = [

| `Consistent of alg_thy'

| `Inconsistent

]

As an optimization, we remember when a theory is consistent or not.

type disj_thy = cached_branches

A disjoint theory is the join of a list of cached_branch. We do not need to * remember the common ancestor of these branches (as an algebraic theory), but only * the atomic cofibrations labeling the path from the common ancestor to each branch.

val dissect_cofibrations : cof list -> branches

This is to dissect the meet of a list of cofibrations into a list of branches. * * Possible further optimizations: * 1. Should we use Cof.reduce to massage the cofibrations first? * 2. Should we eagerly factor out common cofibrations to facilitate the refactoring * steps later on? (This does not seem to be helpful in preliminary experiments.)

module Alg : sig ... end

module Disj : sig ... end