Module Core.Namespace

type +'a t

val empty : 'a t

val add : Ident.t -> 'a -> 'a t -> 'a t

Associate a symbol with a possibly qualified identifier in a namespace.

val nest : ( Stdlib.Format.formatter -> 'a -> unit ) -> [< `Print of string option ] Yuujinchou.Pattern.t -> 'a t -> 'a t -> 'a t

Transform and add a namespace into another namespace.

val find : Ident.t -> 'a t -> 'a option

Look up the symbol associated with an possibly qualified identifier in a namespace.