Module Core.Quote

The purpose of this module is to transform semantic objects into syntactic objects as efficiently as possible; only user-specified top-level definitions will be unfolded, in accordance with the veil (see Monads.QuM.read_veil).

module D := CodeUnit.Domain

module S := CodeUnit.Syntax

val quote_con : CodeUnit.Domain.tp -> CodeUnit.Domain.con -> CodeUnit.Syntax.t Monads.quote

val quote_tp : CodeUnit.Domain.tp -> CodeUnit.Syntax.tp Monads.quote

val quote_sign : CodeUnit.Domain.sign -> CodeUnit.Syntax.sign Monads.quote

val quote_cut : CodeUnit.Domain.cut -> CodeUnit.Syntax.t Monads.quote

val quote_cof : CodeUnit.Domain.cof -> CodeUnit.Syntax.t Monads.quote

val quote_dim : CodeUnit.Domain.dim -> CodeUnit.Syntax.t Monads.quote