Module CofThy.Disj

type t = disj_thy

val envelop_alg' : alg_thy' -> disj_thy

val envelope_alg : [< `Consistent of alg_thy' | `Inconsistent ] -> disj_thy

val empty : t

val consistency : 'a list -> [> `Consistent | `Inconsistent ]

val refactor_branches : (Alg.t' * (VarSet.t * eq list)) list -> t

refactor_branches attempts to identify common parts of the branches * and shrink the labels. Recall that we do not keep the common ancestor * but the paths (as a collection of atomic cofibrations) from the common * ancestor, and thus what are changed here are the paths. * * This optimization seems to be expensive, but it seems to help after * we switched from the persistant tables (using Hashtbl) to Map.

val split : t -> cof list -> t

split thy cofs adds to the theory thy the conjunction of a list of cofibrations cofs * and calculate the branches accordingly. This is similar to Alg.split in the spirit but * differs in detail.

val assume : t -> cof list -> t

assume thy cofs is the same as split thy cofs except that it further refactors the * branches to optimize future searching.

val test_sequent : t -> cof list -> (Cubical.Dim.dim, int) Cubical.Cof.cof -> bool

val left_invert : zero:'a -> seq:(([> `Consistent of 'c ] as 'b -> 'a) -> [> `Consistent of 'c ] list -> 'a) -> ('c * 'd) list -> ('b -> 'a) -> 'a